requirements.txt
Django==3.2.3
Pillow
django-allauth
djangorestframework
django-cors-headers
markdown
django-filter
graphene-django
